Add upload and file display to upload page
Add the capability to select a file and upload it to the submission folder

Add the display which lists the files that are available and allows for
manipulation of those files. Add the handler for the "Delete" links
which removes the file from the display and from the submission.

Add the logic which determines if the questions at the end have been
satisfactorily answered.