Add signin/signout capability
Add the ability to sign in and sign out of the Girder instance
from the technical journal pages.

Add additional menus for when a user is signed in or is an
administrator.

Switch from a generic import in the template files to a full
Girder view for the menu bar.