Redo URLS and add Journal and Issue manipulation pages
Add the code necessary for the admin user to see the list of
issues within the default collection and to add/edit the issues.

Add a separate page to add a new Journal (collection) to the Girder instance

Update the routes to hold the pages necessary and switch to a more
appropriate URL structure.