Add Help pages
Add an administration page which uses the tinymce library to generate the text for
the Journal help pages.  Add pages for the FAQ, About, and main help pages.

Add the webpack capability to pull in the TinyMCE and use it to generate 3 very simple
in-line editors for the page content