Update Journal creation and handling
Add a separate page to allow the main instance to show different journals
other than the default one.
Update the administration page for the journals to list all available
journals and to allow administration on all at once.

Add a new API for capturing and returning all collections which have the
string "__journal__" in the description to allow for different collections
to exist without needing to be reachable from the OTJ interface.