Update folder creation process
As a new submission is generated in the Girder Journal, start a folder
in the user's home directory that is moved to the journal after "finalizing"
the submission and sending it for approval.

Add a new API which allows an administrator to approve it.

Utilize the "curation" plugin to hide the fledgeling plugin.