Fix license metadata for submission
Change from arbitrary value to plain text for license values.
Ensure that approved submissions put metadata in for the folder that
is being approved.  Fix a small spelling mistake.

Add first category for filtering

Add a simple hard-coded category to the main page to filter over
submissions that have code or have testing code.

Add second category and fix searching

Add a second "licensing" category and fix the search to allow the
searching to correctly "OR" between objects in each category and "AND"
between the categories themselves.

Ensure the search dictionary captures the other query values(authors
institution...) such that those searches continue to work.

Add customizable Categories admin and display.

Add a set of pages and APIs  which allows a user to create
customized lists of filters to be displayed on the index page.

Generate "tags" as part of the setting to allow grabbing of types
of values found in the settings model.

Fix style

Fix the results of the linting tests for python and JavaScript

Add missing files

Add the files that were overlooked during a previous commit which is
needed for the home page and the customization pages.