Add search building for each path
Add the code which mirrors the OTJ code to build the query to send each
time a new search event is triggered.

Ensure that the search terms clear correctly when the clear button
is pressed.

Ensure that the "Show more results" link behaves as expected by correcting
the startIndexes sent and checking for the length of the overall return.