Add "search retention" to main page
Use an global variable in the "window" object to pass the last
search query between views.  This should be used to ensure the same search
parameters are used when a user returns to the main page after selecting a submission.

This value shouldn't survive a hard refresh or clicking the "Clear" button