Add download links to upload pages
As new files are uploaded, change the name string into a link
which can download the file.  This will be especially helpful for admins
who are approving the submissions.

Git repository links should not exist right away.  Once girder worker has
captured the zip file, the link should appear and be valid.