Statistics page updates
Ensure that the submissions found during the statistics gathering are
only from the default collection given.

Add a loading wheel to hightlight that the page is doing something and not
frozen.

Change how submissions are collected to ensure that duplicates are not found

Ensure that a year could be appended to the URL to jump directly to the
targeted year.